separate hooks function into another file
