+++ web/components/upload/upload.jsx
@@ -1,52 +1,17 @@
 import {
-  useEffect,
-  useState,
-} from 'react';
+  Icon, Upload,
+} from 'antd';
 import { func } from 'prop-types';
-import { isEmpty } from 'lodash';
-import { Upload, Icon, message } from 'antd';
+
+import { useUpload } from '/components/upload/uploadHooks';
 import { withNamespaces } from '/i18n';
-import { postKeywords } from '/services/keywordsService';
-import { parseCsvFile } from '/utils/papaParse';
 
 import style from './upload.scss';
 
 const { Dragger } = Upload;
 
-const useUpload = (t) => {
-  const [uploadConfig, setUploadConfig] = useState();
-  useEffect(() => {
-    let keywordsData = {};
-    setUploadConfig({
-        name: 'file',
-        accept: 'text/csv',
-        showUploadList: false,
-        async beforeUpload(file) {
-          try {
-            keywordsData = await parseCsvFile(file);
-          } catch (error) {
-            message.error(error);
-          }
-        },
-        async customRequest() {
-          if(isEmpty(keywordsData)) return;
-          try {
-            await postKeywords({ keywordsData });
-            message.success(t('uploadSuccess'));
-          } catch (error) {
-            message.error(error.message);
-          }
-        },
-      });
-  }, []);
-
-  return {
-    uploadConfig,
-  };
-};
-
 const UploadPage = ({ t }) => {
-  const { uploadConfig } = useUpload(t);
+  const { uploadConfig } = useUpload({ t });
   return (
     <div className={style.container}>
       <h3>{t('draggerTitle')}</h3>
@@ -60,7 +25,7 @@ const UploadPage = ({ t }) => {
         </p>
       </Dragger>
     </div>
-  )
+  );
 };
 
 UploadPage.propTypes = {
